Add Home Row Mods Layer
Add HRM Layer that can be toggled on and off in KBD_CTL Layer on right shift since it is a "shifted" version of the base layer because it reacts to tap vs hold on regular keys.
Move Toggle EXT from right Shift to Slash, since it is used mostly for navigating code.
Add CTLS_C to HRM layer to hold both ctl and shift when c is held. This is helpful when selecting text a word at a time.
Change MY_GRV to plain grv on EXT layer since the EXT layer will be used more as a momentary layer most of the time