Add DV LayerLock
While trying to use layer lock keys, noticed that the key only really looks at the top most layer. However, that may not be the layer that we want to lock/unlock. Made a slightly modified version of the library to see if I could get the functionality I want.

DV Layer lock takes into account the layer from which the key was pressed and only acts on it, rather than the top most layer.

Also considers Toggled layers to be locked.

On QK_LLCK, blink the space bar

Currently re-using QK_LLCK key code, but could be changed if needed