Make EXT_LYR act more like FN Layer
instead of treating the EXT_LYR, use it more like a Win_FN Layer that van be re-used by both Base layer and HMR layers. Now EXT has defined the number row as FN keys and arrows now control volume. Removed the ability to swap FN keys from this layer, that feature is still available in the KBD control Layer.

This is getting closer to a stable keymap / layers.